Changed int volume to float to allow volume control

+++ src/ofxGstStandaloneUtils.h
@@ -40,7 +40,7 @@ public:
 	bool  	getIsMovieDone();
 
 	void 	setPosition(float pct);
-	void 	setVolume(int volume);
+	void 	setVolume(float volume);
 	void 	setLoopState(ofLoopType state);
 	int		getLoopState(){return loopMode;}
 	void 	setSpeed(float speed);
