+++ src/ofxGstStandaloneVideoPlayer.h
@@ -43,7 +43,7 @@ public:
 	bool  	getIsMovieDone();
 
 	void 	setPosition(float pct);
-	void 	setVolume(int volume);
+	void 	setVolume(float volume);
 	void 	setLoopState(ofLoopType state);
 	ofLoopType		getLoopState();
 	void 	setSpeed(float speed);
